Move trapProcessExit to its own file
