+++ lib/unexpected-snapshot-update.js
@@ -1,5 +1,6 @@
 const createInspector = require('./createInspector');
 const applyFixes = require('./applyFixes');
+const trapProcessExit = require('./trapProcessExit');
 
 // Make sure that these are the same instances even if the plugin is installed into different expects
 // in different files:
@@ -20,20 +21,7 @@ function ensureAfterHookIsRegistered() {
     const fileNames = Object.keys(fixedSourceTextByFileName);
 
     // Prevent the test runner from forcefully exiting while we're prompting, but save the call args for replaying later:
-    const originalProcessExit = process.exit;
-    const processExitArgs = [];
-    process.exit = function(...args) {
-      processExitArgs.push(args);
-    };
-
-    // Replay process.exit calls that happened while we were prompting and writing files:
-    // eslint-disable-next-line no-inner-declarations
-    function replayAndRestoreProcessExit() {
-      process.exit = originalProcessExit;
-      for (const args of processExitArgs) {
-        process.exit(...args);
-      }
-    }
+    const replayAndRestoreProcessExit = trapProcessExit();
 
     // Give the test runner a chance to complete its output before we prompt:
     setImmediate(async function() {
